Midterm/requirements.txt
spacy~=3.1.0
numpy~=1.21.3
pandas~=1.3.4
gensim~=4.0.1
scikit-learn~=1.0.1
matplotlib~=3.4.3
seaborn~=0.11.2
beautifulsoup4~=4.10.0
requests~=2.26.0
unidecode~=1.3.2
python-dateutil~=2.8.2
wordcloud~=1.8.1